move screenspacemeasurementtool from kinect client into separate class
