+++ source/kinect_client.cpp
@@ -1,4 +1,5 @@
 #include <CMDParser.h>
+#include <screen_space_measurement_tool.hpp>
 
 /// c++ includes
 #include <GL/glew.h>
@@ -16,10 +17,8 @@
 #include <NetKinectArray.h>
 #include <KinectCalibrationFile.h>
 #include <Statistics.h>
-#include <gl_util.h>
 #include <GlPrimitives.h>
-#include <gloostMath.h>
-#include <gloostHelper.h>
+
 
 /// general setup
 unsigned int g_screenWidth  = 1280;
@@ -52,168 +51,8 @@ void idle(void);
 std::vector<kinect::KinectSurfaceV3* > g_ksV3;// 4
 unsigned g_ks_mode = 0;
 
-
-
-struct pickpos{
-public:
-  pickpos(unsigned a, unsigned b)
-    : x(a),
-      y(b)
-  {}
-  unsigned x;
-  unsigned y;
-
-};
-
-gloost::Point3
-multH(gloost::Matrix const& m,
-      gloost::Point3 const& p )
-{
-  gloost::Point3 tmp(m[0]*p[0] + m[4]*p[1] + m[8] *p[2] + m[12] *p[3],
-		     m[1]*p[0] + m[5]*p[1] + m[9] *p[2] + m[13] *p[3],
-		     m[2]*p[0] + m[6]*p[1] + m[10]*p[2] + m[14] *p[3]);
-  tmp[3] =           m[3]*p[0] + m[7]*p[1] + m[11]*p[2] + m[15] *p[3];
-  return tmp;
-}
-
-
-class LineSegmentRenderer{
-public:
-  LineSegmentRenderer()
-  {}
-  ~LineSegmentRenderer()
-  {}
-
-  void draw(std::vector<gloost::Point3> p){
-    glPushAttrib(GL_ALL_ATTRIB_BITS);
-    glDisable(GL_DEPTH_TEST);
-    glDisable(GL_LIGHTING);
-    glColor3f(0.0,1.0,0.0);
-    glPointSize(3.0);
-
-    glBegin(GL_POINTS);
-    for(unsigned i = 0; i < p.size(); ++i){
-      glVertex3fv(p[i].data());
-    }
-    glEnd();
-    glLineWidth(2.0);
-    glBegin(GL_LINE_STRIP);
-    for(unsigned i = 0; i < p.size(); ++i){
-      glVertex3fv(p[i].data());
-    }
-    glEnd();
-
-
-    glPopAttrib();
-  }
-
-private:
-
-};
-
-LineSegmentRenderer* g_lsr = 0;
-
-class ScreenSpaceMeasureTool{
-
-public:
-
-  ScreenSpaceMeasureTool(gloost::Camera* cam, unsigned w, unsigned h)
-    : m_cam(cam),
-      m_w(w),
-      m_h(h),
-      m_pp(),
-      m_mp(),
-      m_modelview()
-  {}
-
-  ~ScreenSpaceMeasureTool()
-  {}
-
-  void resize(unsigned w, unsigned h){
-    m_w = w;
-    m_h = h;
-  }
-
-  void mouse(int button, int state, int mouse_h, int mouse_v){
-    if (button == GLUT_LEFT_BUTTON && state == GLUT_DOWN) {
-      m_pp.clear();
-      pickpos pp(mouse_h, mouse_v);
-      m_pp.push_back(pp);
-
-
-      float d = pickDepth(m_pp[0]);
-
-      // calculate img_to_eye for this view
-      gloost::Matrix viewport_translate;
-      viewport_translate.setIdentity();
-      viewport_translate.setTranslate(1.0,1.0,1.0);
-      gloost::Matrix viewport_scale;
-      viewport_scale.setIdentity();
-      viewport_scale.setScale(m_w * 0.5, m_h * 0.5, 0.5f);
-      gloost::Matrix image_to_eye =  viewport_scale * viewport_translate * m_cam->getProjectionMatrix() * m_modelview;
-      image_to_eye.invert();
-      gloost::Point3 pos_img(m_pp[0].x + 0.5, m_pp[0].y + 0.5, d);
-      gloost::Point3 pos_es = multH(image_to_eye, pos_img);
-      std::cerr << pos_es[3] << std::endl;
-      pos_es[0] /= pos_es[3];
-      pos_es[1] /= pos_es[3];
-      pos_es[2] /= pos_es[3];
-      pos_es[3] /= pos_es[3];
-#if 0      
-      std::cerr << m_cam->getProjectionMatrix() << std::endl;
-      std::cerr << m_pp[0].x << " " << m_pp[0].y << ": " << d << " 3D pos: " << pos_es << std::endl;
-#endif
-
-      m_mp.push_back(pos_es);
-
-    }
-    else if(button == GLUT_RIGHT_BUTTON && state == GLUT_DOWN){
-      m_mp.pop_back();
-    }
-    else if(button == GLUT_MIDDLE_BUTTON && state == GLUT_DOWN){
-      m_mp.clear();
-    }
-  }
-
-  float measure(){
-    const size_t mp_count = m_mp.size();
-    if(mp_count > 1){
-      return (m_mp[mp_count - 1] - m_mp[mp_count - 2]).length();
-    }
-
-    return -1.0;
-  }
-
-  std::vector<gloost::Point3>& getMeasurePoints(){
-    return m_mp;
-  }
-
-  void setModelView(gloost::Matrix& mv){
-    m_modelview = mv;
-  }
-
-private:
-
-  float pickDepth(const pickpos& pp){
-    float d;
-    glReadPixels( pp.x, pp.y, 1, 1, GL_DEPTH_COMPONENT, GL_FLOAT, &d);
-    return d;
-  }
-
-  gloost::Camera*      m_cam;
-  unsigned             m_w;
-  unsigned             m_h;
-
-  std::vector<pickpos> m_pp;
-  std::vector<gloost::Point3> m_mp;
-  gloost::Matrix m_modelview;
-};
-
 bool g_picking = false;
 ScreenSpaceMeasureTool* g_ssmt = 0;
-
-////////////////////////////////////////////////////////////////////////////////
-
 
 void init(std::vector<std::string> args){
 
@@ -227,8 +66,6 @@ void init(std::vector<std::string> args){
   g_ftw = new mvt::FourTiledWindow(g_screenWidth, g_screenHeight);
   g_stats = new mvt::Statistics;
   g_ssmt = new ScreenSpaceMeasureTool(g_camera, g_screenWidth, g_screenHeight);
-
-  g_lsr = new LineSegmentRenderer;
 
   for(unsigned i = 0; i < args.size(); ++i){
     const std::string ext(args[i].substr(args[i].find_last_of(".") + 1));
@@ -418,7 +255,7 @@ void draw3d(void)
   else{
     g_stats->setInfoSlot("navigation mode", 1);
   }
-  g_lsr->draw(g_ssmt->getMeasurePoints());
+  mvt::GlPrimitives::get()->drawLineSegments(g_ssmt->getMeasurePoints());
 
   if(g_reference && g_ksV3.size()){
     glPushAttrib(GL_ALL_ATTRIB_BITS);
@@ -493,7 +330,6 @@ void cleanup(){
   delete g_navi;
   delete g_ftw;
   delete g_ssmt;
-  delete g_lsr;
 }
 
 
@@ -615,12 +451,6 @@ void key(unsigned char key, int x, int y)
   case 't':
     for(unsigned i = 0; i < g_ksV3.size(); ++i){
       g_ksV3[i]->lookup = (int) !g_ksV3[i]->lookup;
-    }
-    break;
-
-  case 'i':
-    for(unsigned i = 0; i < g_ksV3.size(); ++i){
-      g_ksV3[i]->switchCalibVolume();
     }
     break;
   case'f':
